+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## 2.4.1 [unreleased]
 
+* [core] add missing tableName properties in model constructors
+* [space] improve todolist sql resquests
 * [helpdesk] close connection in anycase
 * [core] fix TODO in corespaceadmin, do not get all bookings, just count them
 * [core] if module is inactive, fix authorization (isUserMenuSpaceAuthorized)
 * [core] disable timezone, impacts existing bookings recorded as UTC
+* [booking] fix calendar when area has no resource
 
 ## 2.4.0